Consolidating Code, Small Edits
Continued to work on main as well as turn the entire HiLo game into a function instead of a separate file.
Added a dictionary to name and display descriptions of each of the playable games.
Added an input call to allow user to enter game they want to play and exit program if nothing is entered.